+++ functions.cmake
@@ -1,10 +1,11 @@
 # Create module library
 function(make_lib module srcs link_libs inc_dirs compile_defs)
-    add_library(${module} INTERFACE ${srcs})
-    target_compile_definitions(${module} INTERFACE ${compile_defs})
-    target_link_libraries(${module} INTERFACE ${link_libs})
+    add_library(${module} ${srcs})
+    ament_target_dependencies(${module} PUBLIC ${ROS_DEPS})
+    target_compile_definitions(${module} PUBLIC ${compile_defs})
+    target_link_libraries(${module} PUBLIC ${link_libs})
     target_include_directories(
-        ${module} INTERFACE
+        ${module} PUBLIC
         ${CMAKE_CURRENT_LIST_DIR}/inc
         ${CMAKE_SOURCE_DIR}/lib
         ${inc_dirs}
@@ -12,26 +13,8 @@ function(make_lib module srcs link_libs inc_dirs compile_defs)
     add_dependencies(${module} ${AUTOGEN_TARGETS})
 endfunction()
 
-# Create lib non-ROS executable
-function(make_exe module srcs link_libs inc_dirs ${compile_defs})
-    set(bin_module bin_${module})
-    add_executable(${bin_module} ${srcs})
-    target_compile_definitions(${bin_module} PUBLIC ${compile_defs})
-    target_link_libraries(${bin_module} PUBLIC ${link_libs})
-    target_include_directories(
-        ${bin_module} PUBLIC
-        ${CMAKE_CURRENT_LIST_DIR}/inc
-        ${CMAKE_SOURCE_DIR}/lib
-        ${inc_dirs}
-    )
-    add_dependencies(${bin_module} ${AUTOGEN_TARGETS})
-    install(TARGETS ${bin_module} DESTINATION lib/${PROJECT_NAME})
-    # Rename the output binary to just be the module name
-    set_target_properties(${bin_module} PROPERTIES OUTPUT_NAME ${module})
-endfunction()
-
 # Create project module ROS executable
-function(make_ros_exe module srcs link_libs inc_dirs ${compile_defs})
+function(make_exe module srcs link_libs inc_dirs ${compile_defs})
     set(bin_module bin_${module})
     add_executable(${bin_module} ${srcs})
     target_compile_definitions(${bin_module} PUBLIC ${compile_defs})
