+++ lib/sailbot_db/inc/sailbot_db.h
@@ -141,7 +141,7 @@ public:
     * @param db_name          name of desired database
     * @param mongodb_conn_str URL for mongodb database (ex. mongodb://localhost:27017)
     */
-    explicit SailbotDB(const std::string & db_name, const std::string & mongodb_conn_str);
+    SailbotDB(const std::string & db_name, const std::string & mongodb_conn_str);
 
     /**
      * @brief Format and print a document in the DB
