+++ lib/sailbot_db/test/test_sailbot_db.cpp
@@ -1,0 +1,43 @@
+#include <gtest/gtest.h>
+
+#include "sailbot_db.h"
+#include "util_db.h"
+
+using Polaris::Sensors;
+
+static std::random_device g_rd        = std::random_device();  // random number sampler
+static uint32_t           g_rand_seed = g_rd();                // seed used for random number generation
+static std::mt19937       g_mt(g_rand_seed);                   // initialize random number generator with seed
+// static std::shared_ptr<UtilDB> g_test_db =
+//   UtilDB::initUtilDB("test", MONGODB_CONN_STR, std::make_shared<std::mt19937>(g_mt));
+static std::shared_ptr<UtilDB> g_test_db =
+  std::make_shared<UtilDB>("test", MONGODB_CONN_STR, std::make_shared<std::mt19937>(g_mt));
+class TestSailbotDB : public ::testing::Test
+{
+protected:
+    TestSailbotDB() { g_test_db->cleanDB(); }
+    ~TestSailbotDB() {}
+};
+
+/**
+ * @brief Check that MongoDB is running
+ */
+TEST_F(TestSailbotDB, TestConnection)
+{
+    ASSERT_TRUE(g_test_db->testConnection()) << "MongoDB not running - remember to connect!";
+}
+
+/**
+ * @brief Write random sensor data to the TestDB - read and verify said data
+ */
+TEST_F(TestSailbotDB, TestStoreSensors)
+{
+    SCOPED_TRACE("Seed: " + std::to_string(g_rand_seed));  // Print seed on any failure
+    auto [rand_sensors, rand_info] = g_test_db->genRandData();
+    ASSERT_TRUE(g_test_db->storeNewSensors(rand_sensors, rand_info));
+
+    std::array<Sensors, 1>                expected_sensors = {rand_sensors};
+    std::array<SailbotDB::RcvdMsgInfo, 1> expected_info    = {rand_info};
+
+    g_test_db->verifyDBWrite(expected_sensors, expected_info);
+}
