+++ projects/mock_ais/CMakeLists.txt
@@ -10,18 +10,16 @@ set(inc_dirs
 set(compile_defs
 )
 
-# Create module library
 set(srcs
     ${CMAKE_CURRENT_LIST_DIR}/src/mock_ais.cpp
 )
-make_lib(${module} "${srcs}" "${link_libs}" "${inc_dirs}" "${compile_defs}")
 
 # Create module ROS executable
 set(bin_srcs
     ${srcs}
     ${CMAKE_CURRENT_LIST_DIR}/src/mock_ais_ros_intf.cpp
 )
-make_ros_exe(${module} "${bin_srcs}" "${link_libs}" "${inc_dirs}" "${compile_defs}")
+make_exe(${module} "${bin_srcs}" "${link_libs}" "${inc_dirs}" "${compile_defs}")
 
 # Create unit test
 set(test_srcs
